float32

+++ algorithm/relu.go
@@ -1,13 +1,13 @@
 package algorithm
 
-func Relu(x float64) float64 {
+func Relu(x float32) float32 {
 	if x > 0 {
 		return x
 	}
 	return 0
 }
 
-func ReluDerivative(x float64) float64 {
+func ReluDerivative(x float32) float32 {
 	if x > 0 {
 		return 1
 	}
