+++ data_loader/ollama/gguf_parser/gguf/ggufmeta/main.go
@@ -43,7 +43,7 @@ func main() {
 
 		switch vv := v.(type) {
 		// Scalars:
-		case uint8, int8, uint16, int16, uint32, int32, float32, bool, string, uint64, int64, float64, fmt.Stringer:
+		case uint8, int8, uint16, int16, uint32, int32, float32, bool, string, uint64, int64, float32, fmt.Stringer:
 			print(k, vv)
 
 		// Arrays:
@@ -80,7 +80,7 @@ func main() {
 		case []int64:
 			printArray(k, vv)
 
-		case []float64:
+		case []float32:
 			printArray(k, vv)
 
 		default:
